Revert "Aldaiar"
