--- .gitignore
@@ -1,4 +1,0 @@
-.idea
-__pycache__
-env
-.env
