--- requirements.txt
@@ -1,2 +1,0 @@
-psycopg2-binary==2.9.3
-python-decouple==3.5
